fix(tooling): schema-audit — address PR #1076 round-1 review
Closes 3 of 5 review items from claude-bot round 1:

## #1 — Delete proposal doc per lifecycle rule

Per `07-documentation.md`: "Completed proposals — verify feature is
documented in `reference/`, then delete." The feature is documented in
`docs/reference/tooling/schema-audit.md`; the proposal doc was scaffolding
that captured the council-pass design synthesis. The 4-council-model
history is preserved in the PR description for the merge commit.

Also updates references in `schema-audit.ts` docstring and the reference
doc to drop now-broken links.

## #2 — Strip temporal marker from `schema-audit-parser.ts` header

Removed "milestone-1" reference in the parser file header. Test file
headers had been replaced wholesale by the test-split below; no other
temporal markers remained.

## #3 — Fix `.int.test.ts` exclusion in source-globbing

The previous `glob.replace('*.ts', '*.test.ts')` only matched the FIRST
occurrence of `*.ts` in the glob, so `services/**/*.test.ts` was
produced from `services/**/*.ts` — but `*.int.test.ts` was never
generated as an exclusion. Integration test fixture sites (which create
test users etc.) would have inflated write-site counts.

Replaced with a regex-anchored replace + two exclusion patterns:

```typescript
`!${repoRoot}/${glob.replace(/\*\.ts$/, '*.test.ts')}`,
`!${repoRoot}/${glob.replace(/\*\.ts$/, '*.int.test.ts')}`,
```

## Structure-test compliance (the actual CI red)

The CI `test` check was red because `structure.test.ts` requires every
source file to have a colocated `.test.ts`. My milestone-2 file split
created 5 source files without per-file tests. Split the existing
`schema-audit.test.ts` into 7 colocated test files:

- `schema-audit-parser.test.ts` — schema-parsing tests
- `schema-audit-reads.test.ts` — Recipe Primary tests
- `schema-audit-writes.test.ts` — write-site classification (+ new test
  for `?? null` / `|| null` syntax classification)
- `schema-audit-findings.test.ts` — Recipes Primary/Secondary/Tertiary
- `schema-audit-suppression.test.ts` — audit.config load/validate/apply
- `schema-audit-report.test.ts` — markdown-renderer smoke (stdout capture)
- `schema-audit.test.ts` — entry-point e2e smoke (runSchemaAudit against
  a synthetic project tree, both markdown and JSON output modes)

Added `repoRoot` option to `runSchemaAudit` to support tests without
`process.chdir()` (vitest worker threads don't allow it).

## Tests

747 package tests pass (+6 over milestone-3: the entry-point smokes,
the report renderer, and one new `?? null` write-classification test).

## Deferred to backlog (claude-bot items #4, #5)

- #4 Triple `Project` parse cost — one-shot tool, tolerable
- #5 `withTempDir` async-safety — only matters once `.ts` config tests
  land, which they haven't yet
